requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.3.5
alembic==1.13.3

# Storage
minio==7.2.9

# ML Models
torch==2.4.1
torchvision==0.19.1
open-clip-torch==2.26.1
ultralytics==8.3.2
openai-whisper==20240930

# Video processing
yt-dlp>=2024.8.6

# Configuration
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Apify
apify-client>=1.8.0

# HTTP
httpx==0.27.2
aiohttp==3.10.8

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-httpx==0.30.0

# Enrichment
langdetect==1.0.9
easyocr==1.7.2
vaderSentiment==3.3.2

# GDPR / LLM
langchain==0.3.14
langchain-google-genai==2.0.8

# Utilities
numpy==1.26.4
Pillow==10.4.0
tqdm==4.66.5
setuptools>=70.0.0,<71
